add proper function return types
