+++ frontend/src/components/ScheduledEventsTableCard/index.ts
@@ -1,1 +1,2 @@
 export * from './ScheduledEventsTableCard';
+export * from './utils/fetchData';
